refactor: split FS23 and ADC logic
